requirements.txt
PySide6
matplotlib
numpy
pyusb
